evaluate/requirements.txt
# ──────────────────────────────────────────────
# Core
# ──────────────────────────────────────────────
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.2.0
langchain-text-splitters>=0.3.0

# ──────────────────────────────────────────────
# Vector Store
# ──────────────────────────────────────────────
chromadb>=0.5.0

# ──────────────────────────────────────────────
# Dataset
# ──────────────────────────────────────────────
datasets>=3.0.0

# ──────────────────────────────────────────────
# Evaluation
# ──────────────────────────────────────────────
rouge-score>=0.1.2
nltk>=3.9

# ──────────────────────────────────────────────
# Utilities
# ──────────────────────────────────────────────
python-dotenv>=1.0.0
tqdm>=4.66.0
pandas>=2.2.0
rich>=13.7.0

# ──────────────────────────────────────────────
# GraphRAG
# ──────────────────────────────────────────────
networkx>=3.2.0
